+++ packages/react/src/index.ts
@@ -25,3 +25,8 @@ export type { Theme, ThemeOverrides } from "./design-system/index";
 
 // at the moment we'll re-export everything from the evm package
 export * from "./evm";
+
+// locales
+export { en } from "./evm/locales/en";
+export { ja } from "./evm/locales/ja";
+export { es } from "./evm/locales/es";