+++ packages/react-native-compat/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @thirdweb-dev/react-native-compat
 
+## 0.4.1
+
+## 0.4.0
+
+## 0.3.16
+
 ## 0.3.15
 
 ## 0.3.14